requirements.txt
pandas
plotly
geopandas
leafmap>=0.49.2
ipyleaflet
maplibre
rasterio
rioxarray
solara
widgetsnbextension
geoai-py